requirements.txt
Pillow==8.4.0
pyTelegramBotAPI==4.1.1
python_dateutil==2.8.2
Requests==2.26.0
